refactor: move error codes from types into errors file
