+++ src/apps/backend/modules/application/errors.py
@@ -1,5 +1,17 @@
-from modules.application.types import WorkerErrorCode
+from dataclasses import dataclass
+
 from modules.error.custom_errors import AppError
+
+
+@dataclass(frozen=True)
+class WorkerErrorCode:
+    WORKER_CLIENT_CONNECTION_ERROR: str = "WORKER_ERR_01"
+    WORKER_CLASS_INVALID: str = "WORKER_ERR_02"
+    WORKER_WITH_ID_NOT_FOUND: str = "WORKER_ERR_03"
+    WORKER_START_ERROR: str = "WORKER_ERR_04"
+    WORKER_ALREADY_COMPLETED: str = "WORKER_ERR_05"
+    WORKER_ALREADY_CANCELLED: str = "WORKER_ERR_06"
+    WORKER_ALREADY_TERMINATED: str = "WORKER_ERR_07"
 
 
 class WorkerClientConnectionError(AppError):
@@ -17,8 +29,7 @@ class WorkerIdNotFoundError(AppError):
         super().__init__(
             code=WorkerErrorCode.WORKER_WITH_ID_NOT_FOUND,
             http_status_code=404,
-            message=f"Worker with given id: {worker_id} not found. "
-            f"Verify the ID of the worker and try again.",
+            message=f"Worker with given id: {worker_id} not found. " f"Verify the ID of the worker and try again.",
         )
 
 
@@ -47,8 +58,7 @@ class WorkerAlreadyCancelledError(AppError):
         super().__init__(
             code=WorkerErrorCode.WORKER_ALREADY_CANCELLED,
             http_status_code=400,
-            message=f"Worker with id: {worker_id} has already been cancelled."
-            f"Verify the worker ID and try again.",
+            message=f"Worker with id: {worker_id} has already been cancelled." f"Verify the worker ID and try again.",
         )
 
 
@@ -57,6 +67,5 @@ class WorkerAlreadyTerminatedError(AppError):
         super().__init__(
             code=WorkerErrorCode.WORKER_ALREADY_TERMINATED,
             http_status_code=400,
-            message=f"Worker with id: {worker_id} has already been terminated. "
-            f"Verify the worker ID and try again.",
+            message=f"Worker with id: {worker_id} has already been terminated. " f"Verify the worker ID and try again.",
         )
