+++ src/apps/backend/modules/application/types.py
@@ -40,14 +40,3 @@ class Worker:
     close_time: Optional[datetime]
     task_queue: str
     worker_type: str
-
-
-@dataclass(frozen=True)
-class WorkerErrorCode:
-    WORKER_CLIENT_CONNECTION_ERROR: str = "WORKER_ERR_01"
-    WORKER_CLASS_INVALID: str = "WORKER_ERR_02"
-    WORKER_WITH_ID_NOT_FOUND: str = "WORKER_ERR_03"
-    WORKER_START_ERROR: str = "WORKER_ERR_04"
-    WORKER_ALREADY_COMPLETED: str = "WORKER_ERR_05"
-    WORKER_ALREADY_CANCELLED: str = "WORKER_ERR_06"
-    WORKER_ALREADY_TERMINATED: str = "WORKER_ERR_07"
